move batch id
