+++ src/sr2silo/vpipe/__init__.py
@@ -2,3 +2,9 @@
 
   i.e. extracting metadata from V-Pipe Filenaming Conventions.
   """
+
+from __future__ import annotations
+
+from sr2silo.vpipe.metadata import batch_id_decoder, sample_id_decoder
+
+__all__ = ["sample_id_decoder", "batch_id_decoder"]
